requirements.txt
fastapi
uvicorn[standard]
pymongo
certifi
numpy
sentence-transformers
faiss-cpu
groq
python-dotenv
python-multipart
pydantic
httpx
